% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cpf}
\alias{cpf}
\title{Package-cpf}
\description{
This package contains functions to plot a dataframe containing the total aamount of funds by country
for the EU Cohesion policy 2014/2020. This package Runs also a basic Interface witch shows the EU Member States.
}
