% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_cpf.R
\name{cpfapp}
\alias{cpfapp}
\title{Shiny}
\usage{
cpfapp()
}
\value{
shiny app
}
\description{
Runs a basic Interface witch shows the Member States and the total
amount of funds for the european Social Cohesion policy by country
}
