% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcp.R
\name{plotcp}
\alias{plotcp}
\title{Illustration of Cohesion policy funds}
\usage{
plotcp(data, aes_string)
}
\arguments{
\item{data}{is a data.frame}

\item{aes_string}{are the mapping coordinates}
}
\value{
a ggplot2 object
}
\description{
it creates a horizontal lollipop chart representing the total funds
}
\details{
Creates a lollipop chart of total funds by country
}
\author{
Anna Olena Zhab'yak
}
