% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcp.R
\name{plotcp}
\alias{plotcp}
\title{Illustration of Cohesion policy funds}
\usage{
plotcp(x, y)
}
\arguments{
\item{x}{is a character vector}

\item{y}{is a numeric vector}
}
\description{
it creates a horizontal graph representing the total funds
}
\details{
Creates a plot of total funds by country
}
\author{
Anna Olena Zhab'yak
}
